project lighting stuff added

+++ project/project/raycast.h
@@ -52,7 +52,7 @@ public:
 
 protected:
   virtual float degToRad(int a) { return a*M_PI/180.0;};
-  virtual int FixAng(int a){ if(a>359){ a-=360;} if(a<0){ a+=360;} return a;};
+  virtual float FixAng(float a){ if(a>359){ a-=360;} if(a<0){ a+=360;} return a;};
   virtual float distt(float a, float b)
   {
     return (sqrt(a*a + b*b));
